requirements.txt
playwright==1.49.0
Pillow>=10.0.0
PyYAML>=6.0
